+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.17.1] - 2022-09-10
+### Fixed
+- toaster notification blocking clicking whole bottom of the screen
+- actions on vertical header not having proper spaces and EqualActions not working correctly for vertical header at all
+
 ## [0.17.0] - 2022-07-26
 ### Added
 - `KeyValue` component